ascii_arts nincat: now arts with external commands must include a prefix
Had this idea of doing this a long time ago, just didn't know the best
way to actually do it.

This implementation is a one-call of identifying an art that needs
an external command and running this external command.

This can be implemented for art files specifically instead of art files
and of course, the prefix can be changed for something better and
unique, because if there is a folder before this with the same prefix,it
will actually try to run it. I don't know if grep can run reversely, but
if it can, maybe later I add its option.

Anyways, this took actually some time, but it is fun and not breakable
at all! 100% code coverage!